Backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1

sqlalchemy==2.0.30
psycopg2-binary==2.9.9

alembic==1.13.1

pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1

bcrypt==4.1.3
PyJWT==2.8.0

email-validator==2.1.1
python-multipart==0.0.9